script-breakdown-260307/requirements.txt
PyQt6>=6.5.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
lxml>=5.0.0
anthropic>=0.40.0
openai>=1.50.0
requests>=2.31.0
flask>=3.0.0
ortools>=9.8
